new update

+++ src/todos/todo.service.ts
@@ -3,6 +3,9 @@ import { Todo } from './entity/todo.entity';
 import { CreateTodoDto } from './dtos/craete-todo.dto';
 import { UpdateTodoDto } from './dtos/update-todo.dto';
 import { Tag } from './entity/tag.model';
+import { UploadApiErrorResponse, UploadApiResponse, v2 } from 'cloudinary';
+import toStream = require('buffer-to-stream');
+import { UserEntity } from 'src/auth/entity/user.entity';
 
 @Injectable()
 export class TodoService {
@@ -13,9 +16,14 @@ export class TodoService {
     private readonly tagRepository: typeof Tag,
   ) {}
 
-  async findAll() {
+  async findAll(page: number) {
     try {
-      const todoList = await this.todoRepository.findAll();
+      const todosPerPage = 10;
+      const offset = (page - 1) * todosPerPage;
+      const todoList = await this.todoRepository.findAll({
+        limit: todosPerPage,
+        offset,
+      });
       return {
         status: {
           success: true,
@@ -41,21 +49,77 @@ export class TodoService {
     });
   }
 
-  async createTodo(createTodoDto: CreateTodoDto, createdBy: number) {
-    createTodoDto.createdBy = createdBy;
-
-    const [tag] = await this.tagRepository.findOrCreate({
-      where: { tag_name: createTodoDto.tagName },
+  async findOneTag(id: number) {
+    return await this.tagRepository.findOne({
+      where: {
+        id,
+      },
+      include: [Tag],
     });
+  }
+
+  async findUserTodo(userId: number, page: number) {
+    try {
+      const todosPerPage = 10;
+      const offset = (page - 1) * todosPerPage;
+      const todoList = await this.todoRepository.findAll({
+        where: { userId },
+        limit: todosPerPage,
+        offset,
+      });
+      return {
+        status: {
+          success: true,
+          code: 200,
+          message: 'Get All Data Successfuly',
+        },
+        data: todoList,
+      };
+    } catch (error) {
+      throw new HttpException(
+        'something want wrong!',
+        HttpStatus.SERVICE_UNAVAILABLE,
+      );
+    }
+  }
+
+  async uploadImage(
+    file: Express.Multer.File,
+  ): Promise<UploadApiResponse | UploadApiErrorResponse> {
+    return new Promise((resolve, reject) => {
+      const upload = v2.uploader.upload_stream((error, result) => {
+        if (error) return reject(error);
+        resolve(result);
+      });
+
+      toStream(file.buffer).pipe(upload);
+    });
+  }
+
+  async createTodo(
+    createTodoDto: CreateTodoDto,
+    userId: number,
+    file: Express.Multer.File,
+  ) {
+    let imageUrl = null;
+    imageUrl = await this.uploadImage(file);
+
+    const newTag = await this.tagRepository.build({
+      tag_name: createTodoDto.tagName,
+    });
+
+    await newTag.save();
 
     const newTodo = await this.todoRepository.build({
       title: createTodoDto.title,
       description: createTodoDto.description,
       isCompleted: createTodoDto.isCompleted,
-      tagName: tag.tag_name,
+      tagName: newTag.tag_name,
+      avatar: imageUrl['secure_url'],
+      userId: userId,
     });
 
-    newTodo.tag = tag;
+    newTodo.tag = newTag;
 
     await newTodo.save();
 
@@ -70,9 +134,9 @@ export class TodoService {
   }
 
   async updateTodo(
+    currentUser: UserEntity,
     id: number,
     updateTodoDto: UpdateTodoDto,
-    updatedBy: number,
   ) {
     const todoUpdate = await this.findOne(id);
 
@@ -83,11 +147,30 @@ export class TodoService {
       );
     }
 
+    if (currentUser.id !== id) {
+      throw new HttpException(
+        'You are not authorized to update this account.',
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
+
+    if (updateTodoDto.tagName) {
+      const tag = await this.tagRepository.findOne({
+        where: { id },
+      });
+
+      await tag.update({ tag_name: updateTodoDto.tagName });
+
+      await this.tagRepository.update(tag, {
+        where: { id },
+        returning: true,
+      });
+    }
+
     Object.assign(todoUpdate, updateTodoDto);
 
     await todoUpdate.save();
 
-    updateTodoDto.updatedBy = updatedBy;
     await this.todoRepository.update(updateTodoDto, {
       where: { id },
       returning: true,
@@ -96,7 +179,7 @@ export class TodoService {
     return await this.findOne(id);
   }
 
-  async deleteTodo(id: number) {
+  async deleteTodo(currentUser: UserEntity, id: number) {
     const todoDelete = await this.findOne(id);
 
     if (todoDelete === null) {
@@ -105,6 +188,14 @@ export class TodoService {
         HttpStatus.NOT_FOUND,
       );
     }
+
+    if (currentUser.id !== id) {
+      throw new HttpException(
+        'You are not authorized to update this account.',
+        HttpStatus.UNAUTHORIZED,
+      );
+    }
+
     await this.todoRepository.destroy({ where: { id } });
     await this.tagRepository.destroy({ where: { id } });
